test/variable runs in isolation
